Add new section for newsletter sign-up and change page style
A new div section has been added to index.html to include a form for users to enter their email address to sign up for a newsletter or updates providing interaction with them. The styling adjustments in style.css change the body background color to a darker shade and the h1 text color to gold enhancing page's overall look. It also includes flexbox styling for newly added .container and .building divs to structure the page layout better, the width, height, and border of these divs have been defined as well.